# specs/001-course-assistant-mvp/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DOU-Synapse API",
    "description": "Ders materyaline bağlı, kaynaklı cevap üreten ders asistanının HTTP yüzeyi. Her cevap dayandığı sayfayla döner; dayanak yoksa uç cevap değil gerekçeli ret döndürür. Ders izolasyonu iki katmanda zorlanır: sunucu tarafında üyelik doğrulaması ve aynı oturumda PostgreSQL satır düzeyi güvenlik.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/courses/{course_id}/chat/sessions/{session_id}": {
      "delete": {
        "tags": [
          "privacy"
        ],
        "summary": "Delete Chat Session",
        "operationId": "delete_chat_session_courses__course_id__chat_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDeletionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_courses__course_id__chat_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_ChatMessageOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/chat/sessions": {
      "delete": {
        "tags": [
          "privacy"
        ],
        "summary": "Delete Course Chat History",
        "operationId": "delete_course_chat_history_courses__course_id__chat_sessions_delete",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDeletionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Sessions",
        "description": "Kullanıcının bu dersteki sohbet oturumları. RLS başkasınınkini zaten göstermez.",
        "operationId": "list_sessions_courses__course_id__chat_sessions_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Ders kimliği",
              "title": "Course Id"
            },
            "description": "Ders kimliği"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOut_ChatSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerStatus": {
        "type": "string",
        "enum": [
          "answered",
          "insufficient_context",
          "out_of_scope",
          "budget_exhausted"
        ],
        "title": "AnswerStatus",
        "description": "Cevabın dört meşru sonucu. Üçü 'cevap yok' ama SEBEPLERİ farklı.\n\nKarıştırılmamalı: `insufficient_context` \"materyalde var olabilir ama kanıt\nzayıf\", `out_of_scope` \"bu ders bu konuyu hiç kapsamıyor\". Kullanıcıya\nverilen metin de ölçüm kategorisi de farklıdır (SC-005 yalnız ikincisini ölçer)."
      },
...
      "AssistantAudience": {
        "type": "string",
        "enum": [
          "student",
          "instructor"
        ],
        "title": "AssistantAudience",
        "description": "Server-derived course role used by the assistant; never client-selected."
      },
...
      "ChatDeletionOut": {
        "properties": {
          "deleted_sessions": {
            "type": "integer",
            "title": "Deleted Sessions"
          }
        },
        "type": "object",
        "required": [
          "deleted_sessions"
        ],
        "title": "ChatDeletionOut"
      },
      "ChatFeedbackOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "message_id": {
            "type": "string",
            "format": "uuid",
            "title": "Message Id"
          },
          "rating": {
            "$ref": "#/components/schemas/ChatFeedbackRating"
          },
          "reason": {
            "$ref": "#/components/schemas/ChatFeedbackReason"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "share_with_instructor": {
            "type": "boolean",
            "title": "Share With Instructor"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_id",
          "rating",
          "reason",
          "comment",
          "share_with_instructor",
          "created_at",
          "updated_at"
        ],
        "title": "ChatFeedbackOut"
      },
      "ChatFeedbackRating": {
        "type": "string",
        "enum": [
          "helpful",
          "unhelpful"
        ],
        "title": "ChatFeedbackRating"
      },
      "ChatFeedbackReason": {
        "type": "string",
        "enum": [
          "helpful",
          "inaccurate",
          "irrelevant",
          "citation_problem",
          "too_direct",
          "unsafe",
          "other"
        ],
        "title": "ChatFeedbackReason"
      },
...
      "ChatMessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/ChatRole"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "type": "array",
            "title": "Citations"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "socratic_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SocraticStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatFeedbackOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "ChatMessageOut"
      },
      "ChatMode": {
        "type": "string",
        "enum": [
          "qa",
          "socratic",
          "exam"
        ],
        "title": "ChatMode"
      },
...
      "ChatRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant"
        ],
        "title": "ChatRole",
        "description": "Mesajın sahibi. `contracts.py`'de yok çünkü modüller arası bir tip değil."
      },
      "ChatSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "mode": {
            "$ref": "#/components/schemas/ChatMode"
          },
          "audience": {
            "$ref": "#/components/schemas/AssistantAudience"
          },
          "agent_profile": {
            "type": "string",
            "title": "Agent Profile"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "socratic_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SocraticStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "title": "Created At"
          },
          "updated_at": {
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "mode",
          "audience",
          "agent_profile",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionOut"
      },
...
      "CitationOut": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "claim": {
            "type": "string",
            "title": "Claim",
            "default": ""
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "file_name",
          "location",
          "snippet"
        ],
        "title": "CitationOut",
        "description": "Kullanıcıya gösterilen atıf.\n\n`file_name`, `location` ve `snippet` chunk metadata'sından gelir. Modelin\nmetninden türetilen tek alan `claim`'dir ve o da yalnız açıklayıcıdır —\nhiçbir güvenlik kararı ona bakmaz."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageOut_ChatMessageOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PageOut[ChatMessageOut]"
      },
      "PageOut_ChatSessionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatSessionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PageOut[ChatSessionOut]"
      },
...
      "SocraticStage": {
        "type": "string",
        "enum": [
          "diagnose",
          "nudge",
          "concept_hint",
          "similar_example",
          "explain_with_source"
        ],
        "title": "SocraticStage",
        "description": "Sokratik merdiven. Sıra ANLAMLIDIR ve atlanamaz (spec FR-014).\n\nÖğrenci kendi denemesini yapmadan bir sonraki kademeye geçilmez; bu kural\nstate machine'de zorlanır, prompt'a bırakılmaz."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}